requirements.txt
# Core dependencies
pandas>=2.1.0
numpy>=1.26.0
scikit-learn>=1.3.0
pyarrow>=14.0.0  # For parquet file support

# Data fetching
yfinance>=0.2.32
ccxt>=4.1.0  # For forex/crypto data

# Technical analysis
# Note: We implement our own technical indicators using pandas/numpy
# ta-lib is optional - install separately if needed: sudo apt-get install ta-lib && pip install TA-Lib
# pandas-ta is optional - install if needed: pip install pandas-ta

# Machine Learning & RL
# TensorFlow may have compatibility issues with Python 3.12
# Using compatible versions or alternatives
stable-baselines3>=2.2.0
gym>=0.26.0

# Database
sqlalchemy>=2.0.0
# Note: sqlite3 is built-in to Python, no need to install

# Telegram Bot
python-telegram-bot>=20.0

# Visualization
matplotlib>=3.8.0
plotly>=5.18.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
aiohttp>=3.9.0

# Backtesting
backtrader>=1.9.70

# Note: tensorflow and keras are optional for basic functionality
# If you need full RL capabilities, install separately:
# pip install tensorflow==2.15.0 keras==2.15.0
# Note: TensorFlow 2.15 may not support Python 3.12, consider using Python 3.11

